Improve logging.  Don't bail tests on a failure.

+++ qa/integration5.0snapshot.sh
@@ -27,7 +27,7 @@ fi
 java_version=$(java -version 2>&1 | grep version | sed 's|.* version "\(.*\..*\)\..*_.*"|\1|')
 ubuntu_version=$(grep VERSION_ID /etc/*-release)
 if [[ $java_version < 1.8 ]]; then
-  echo "Install Java 8, please wait"
+  echo "-- `date` Install Java 8, please wait"
   sudo add-apt-repository -y ppa:webupd8team/java &> /dev/null
   sudo apt-get -qq update &> /dev/null
   echo "oracle-java8-installer shared/accepted-oracle-license-v1-1 select true" | sudo /usr/bin/debconf-set-selections &> /dev/null
@@ -35,12 +35,12 @@ if [[ $java_version < 1.8 ]]; then
 fi
 
 java_version=$(java -version 2>&1 | grep version | sed 's|.* version "\(.*\..*\)\..*_.*"|\1|')
-echo "Java version = $java_version"
+echo "-- `date` Java version = $java_version"
 if [[ $java_version < 1.8 ]]; then
   exit 1
 fi
 
-echo "-- install libfontconfig libfreetype6 so Reporting can work on a headless server"
+echo "-- `date` install libfontconfig libfreetype6 so Reporting can work on a headless server"
 sudo apt-get -qq install -y  libfontconfig libfreetype6
 
 . ./setenv.sh
@@ -56,31 +56,56 @@ BASEURL=snapshots.elastic.co
 
 #./cleanup.sh
 
-echo "-- Get the packages"
-echo http://${BASEURL}/download/beats/packetbeat/packetbeat-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE}
-ls packetbeat-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE} || wget -q http://${BASEURL}/download/beats/packetbeat/packetbeat-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE} || exit 1
-ls filebeat-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE}   || wget -q http://${BASEURL}/download/beats/filebeat/filebeat-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE} || exit 1
-ls metricbeat-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE} || wget -q http://${BASEURL}/download/beats/metricbeat/metricbeat-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE} || exit 1
-ls kibana-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE}     || wget -q http://${BASEURL}/download/kibana/kibana-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE} || exit 1
+echo "-- `date` Get the packages"
+if [ ! -f packetbeat-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE} ]; then
+  echo "wget -q http://${BASEURL}/download/beats/packetbeat/packetbeat-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE}"
+  wget -q http://${BASEURL}/download/beats/packetbeat/packetbeat-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE} &
+fi
 
-ls logstash-${VERSION}${SNAPSHOT}.${PACKAGE}      || wget -q http://${BASEURL}/download/logstash/logstash-${VERSION}${SNAPSHOT}.${PACKAGE} || exit 1
-ls elasticsearch-${VERSION}${SNAPSHOT}.${PACKAGE} || wget -q http://${BASEURL}/download/elasticsearch/elasticsearch-${VERSION}${SNAPSHOT}.${PACKAGE} || exit 1
+if [ ! -f filebeat-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE} ]; then
+  echo "wget -q http://${BASEURL}/download/beats/filebeat/filebeat-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE}"
+  wget -q http://${BASEURL}/download/beats/filebeat/filebeat-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE} &
+fi
 
-echo "-- Install packages"
+if [ ! -f metricbeat-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE} ]; then
+  echo "wget -q http://${BASEURL}/download/beats/metricbeat/metricbeat-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE}"
+  wget -q http://${BASEURL}/download/beats/metricbeat/metricbeat-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE} &
+fi
+
+if [ ! -f kibana-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE} ]; then
+  echo "wget -q http://${BASEURL}/download/kibana/kibana-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE}"
+  wget -q http://${BASEURL}/download/kibana/kibana-${VERSION}${SNAPSHOT}${PLATFORM}.${PACKAGE} &
+fi
+
+if [ ! -f logstash-${VERSION}${SNAPSHOT}.${PACKAGE} ]; then
+  echo "wget -q http://${BASEURL}/download/logstash/logstash-${VERSION}${SNAPSHOT}.${PACKAGE}"
+  wget -q http://${BASEURL}/download/logstash/logstash-${VERSION}${SNAPSHOT}.${PACKAGE} &
+fi
+
+if [ ! -f elasticsearch-${VERSION}${SNAPSHOT}.${PACKAGE} ]; then
+  echo "wget -q http://${BASEURL}/download/elasticsearch/elasticsearch-${VERSION}${SNAPSHOT}.${PACKAGE}"
+  wget -q http://${BASEURL}/download/elasticsearch/elasticsearch-${VERSION}${SNAPSHOT}.${PACKAGE} &
+fi
+
+echo "-- `date` Wait for all packages to download"
+wait
+
+echo "-- `date` Install packages"
 ./install_packages.sh || exit 1
 
 # Install Platinum License?
 
-echo "-- Configure beats authenication"
+echo "-- `date` Configure beats authenication"
 ./configure_beats.sh || exit 1
 
-echo "-- Install Elasticsearch X-Pack"
-#                                                                 download/elasticsearch-plugins/x-pack/x-pack-5.0.0-alpha6-SNAPSHOT.zip
+echo "-- `date` Install Elasticsearch X-Pack"
 if [ .$HASH. == .. ]; then
   echo "Getting http://${BASEURL}/download/elasticsearch-plugins/x-pack/x-pack-${VERSION}${SNAPSHOT}.zip"
-  ls x-pack-${VERSION}${SNAPSHOT}.zip || wget -q http://${BASEURL}/download/elasticsearch-plugins/x-pack/x-pack-${VERSION}${SNAPSHOT}.zip || exit 1
-  echo "time sudo /usr/share/elasticsearch/bin/elasticsearch-plugin install -b file:///${QADIR}/x-pack-${VERSION}${SNAPSHOT}.zip"
-  time sudo /usr/share/elasticsearch/bin/elasticsearch-plugin install -b file:///${QADIR}/x-pack-${VERSION}${SNAPSHOT}.zip
+  if [ ! -f -pack-${VERSION}${SNAPSHOT}.zip ]; then
+    wget -q http://${BASEURL}/download/elasticsearch-plugins/x-pack/x-pack-${VERSION}${SNAPSHOT}.zip || exit 1
+    echo "time sudo /usr/share/elasticsearch/bin/elasticsearch-plugin install -b file:///${QADIR}/x-pack-${VERSION}${SNAPSHOT}.zip"
+    time sudo /usr/share/elasticsearch/bin/elasticsearch-plugin install -b file:///${QADIR}/x-pack-${VERSION}${SNAPSHOT}.zip
+  fi
 else
   echo "Staging install using ES_JAVA_OPTS=\"-Des.plugins.staging=$HASH\""
   #wget http://staging.elastic.co/5.0.0-alpha6-b2c88dcc/download/elasticsearch-plugins/x-pack/x-pack-5.0.0-alpha6.zip
@@ -91,7 +116,7 @@ else
   #https://artifacts.elastic.co/download/elasticsearch-plugins/x-pack/x-pack-5.0.0-alpha6.zip
 fi
 
-echo "-- Install Kibana UI Plugins https://${BASEURL}/download/kibana/plugins/x-pack/x-pack-${VERSION}${SNAPSHOT}.zip"
+echo "-- `date` Install Kibana UI Plugins https://${BASEURL}/download/kibana/plugins/x-pack/x-pack-${VERSION}${SNAPSHOT}.zip"
 #wget https://staging.elastic.co/5.0.0-alpha6-b2c88dcc/download/kibana-plugins/x-pack/x-pack-5.0.0-alpha6.zip
 time sudo /usr/share/kibana/bin/kibana-plugin install https://${BASEURL}/download/kibana-plugins/x-pack/x-pack-${VERSION}${SNAPSHOT}.zip | grep -v "^\.$"
 
@@ -100,7 +125,7 @@ time sudo /usr/share/kibana/bin/kibana-plugin install https://${BASEURL}/downloa
 # create user for logstash to connect to Elasticsearch (used in logstash.conf)
 #/usr/share/elasticsearch/bin/x-pack/users useradd $LOGSTASHUSER -r logstash -p $LOGSTASHPWD || exit 1
 
-echo "-- let logstash process read syslog"
+echo "-- `date` let logstash process read syslog"
 #sudo setfacl -m u:logstash:r /var/log/syslog || exit 1 (this VM is not mounted with acls)
 sudo chmod o+r /var/log/syslog
 sudo cp logstash.conf /etc/logstash/conf.d/ || exit 1
@@ -109,30 +134,30 @@ sudo cp logstash.conf /etc/logstash/conf.d/ || exit 1
 
 # curl put watcher trigger data?
 
-echo "-- Set network.host for Elasticsearch so we can access it outside the vagrant machine"
+echo "-- `date` Set network.host for Elasticsearch so we can access it outside the vagrant machine"
 echo "network.host: 0.0.0.0" >> /etc/elasticsearch/elasticsearch.yml
 # Now that we set host, we have to make it more production like
 echo "discovery.zen.minimum_master_nodes: 0" >> /etc/elasticsearch/elasticsearch.yml
 
-echo "-- set jvm min heap size"
+echo "-- `date` set jvm min heap size"
 sed -i 's/-Xms256m/-Xms2g/' /etc/elasticsearch/jvm.options
 
-echo "-- Set network.host for Kibana so we can access it outside the vagrant machine"
+echo "-- `date` Set network.host for Kibana so we can access it outside the vagrant machine"
 echo "server.host: 0.0.0.0" >> /etc/kibana/kibana.yml
 
-echo "-- Start services"
+echo "-- `date` Start services"
 ./start_services.sh || exit 1
 
 # make some packet data
 ping -c 100 www.google.com >/dev/null &
 
-echo "-- Wait for Elasticsearch and Kibana to be ready"
+echo "-- `date` Wait for Elasticsearch and Kibana to be ready"
 ./check.sh
 
-echo "-- Create a Kibana user (iron man)"
+echo "-- `date` Create a Kibana user (iron man)"
 ./create_kibana_user.sh >/dev/null
 
-echo "-- Load Beats index patterns, saves searches, visualizations, and dashboards"
+echo "-- `date` Load Beats index patterns, saves searches, visualizations, and dashboards"
 pushd /usr/share/filebeat/scripts/
 ./import_dashboards -user $ELASTICUSER -pass $ELASTICPWD -url http://${BASEURL}/beats/dashboards/beats-dashboards-${VERSION}.zip
 popd
